feat: split each section of task UI to its own component
